Feira Medieval 2024

+++ src/rtfPrint.js
@@ -6,15 +6,10 @@ const util = require('util');
 novoArquivoRTF = path.resolve(__dirname, './senha.rtf');
 
 async function printRTF(produto, quantidade, preco, total, nPrint) {
-  try {
-    fs.statSync(novoArquivoRTF);
+  if (fs.existsSync(novoArquivoRTF)) {
     fs.unlinkSync(novoArquivoRTF);
-  } catch (error) {
-    if (error.code === 'ENOENT') {
-      console.log('O arquivo não existe, não é necessário excluí-lo.');
-    } else {
-      throw error;
-    }
+  } else {
+    console.log('O arquivo não existe, não é necessário excluí-lo.');
   }
 
   const rtfTemplate = `{\\rtf1\\ansi\\ansicpg1252\\deff0\\nouicompat\\deflang2070{\\fonttbl{\\f0\\fnil\\fcharset0 Curlz MT;}{\\f1\\fnil\\fcharset0 Calibri;}}
